proxies / mirrormanager / download redirect: fix https/http mixing
Fixes https://pagure.io/fedora-infrastructure/issue/9564

download.fedoraproject.org queries mirrormanager for a redirect to a
mirror for the path/request. Before we were just taking any mirror that
mirrormanager had, if it was http or https. This caused requests that
were sent in as https to get a http mirror and error out. So, now we
just redirect http ones to http mirrors and https requests to https
mirrors.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>