copr-be: increase the chroot tmpfs space on builders
Mock allocated 75g before, and this started to be not enough for some
projects (tmpfs resides in ram + swap).  Increasing the size isn't a
huge deal for us since the major subset of builders is x86_64 and those
have pretty large (currently unused) swap disk size.  For aarch64 I had
to add a bit more storage to guarantee that the large tmpfs will fit
the swap (+17%).