ipa/client: Always warn about `fas_client_groups`
The `fas_client_groups` and `ipa_client_shell_groups` variables have a
different format, the former is a comma-separated string, the latter a
list. Nag about it with more detail and regardless of if
`ipa_client_shell_groups` is set, because if it is, then the old
variable gets ignored.

Signed-off-by: Nils Philippsen <nils@redhat.com>