Set noggin default avatar to retro
The default in noggin for the libravatar default avatar (i.e. the avatar
that shows when a user hasnt set their avatar) is an autogenerated
robot.

However, the majority of other applications in Fedora Infra use the
"retro" avatar option. This changes the config of noggin in Fedora
Accounts to use the retro option, rather than the robot one.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>