This pull request sets up a vmhost COPR environment in RDU-CC
First we add in a vmhost_copr playbook which is basically the regular
virthost but meant to limit things for copr admins via rbac.

Second we add in host and group variables which will use the ansible
module for setting up networks. This allows for bridging to work out
of the box and makes sure we know the mac addresses.