Try to get virthosts to use correct postfix set.
For some reason the vmhost-x86-copr systems are not using the correct
main.cf even though we have told the variables which one to use. I am
trying a second method to try and force this, but will have to come up
with a different datacenter definition for these otherwise as I can't
figure out why the code is different.