distgit/pagure: remove the commit ACL for API keys in prod
This ACL turns out to be too confusing to users as it currently
does not work with our OIDC set-up with fedpkg.
Once we'll have figured out how to make both work together or
keep one and remove the other, we can revisit.

Keeping this in staging so we have a place where we can experiment
with this.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>