Fix traceback when the creation date is exactly at 0 ms
nb found that one user was blocking the cronjob from running.
After looking closely, the problem was the creation date is
similar to 2017-02-01 09:10:20+00:00 , so without any dot,
as it was created at the exact microsecond the second started.

The usual format is 2017-02-01 09:10:20.012+00:00, where
split('.') work fine.

Since the traceback stop the whole cronjob, this prevented all
people whose login was after m from having the lifecycle badges.