filter_plugins / fedmsg: try and make staging only talk to staging hosts
I think the only reason this worked in phx2 was because we also had a
firewall rule that was blocking the staging subnet on all prod machines.
Instead of doing that here, lets just make fedmsg only deal with the
staging hosts in staging. The only thing this might break is things that
aren't in staging_friendly that should be.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>