sundries and proxies / staging: switch to new flatpak-indexer
Under the old setup, we ran 'reg' on sundries and then synced that
content (in 3 rsyncs) to proxies.

With this new setup, flatpak-indexer runs in openshift. It has a nfs
volume where it writes it's output too. We rsync just that entire output
directory to the proxies. This includes things like deltas.

This commit should only affect staging, we will roll to production
later.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>