Make fixing home dir ownership a separate playbook
We will have to fix ownership of home directories on all FAS client
hosts when migrating users to IPA because they have different uids
there. This is a manual playbook because running this on fedorapeople
will have to check thousands of directories which is a cost we want to
avoid in ordinary playbook runs.

Also, tag tasks in the reenrollment playbook to allow specifying which
parts get run if necessary.

Signed-off-by: Nils Philippsen <nils@redhat.com>