copr-be-dev: increase quota for one user
Nobody is using devel stack except for Copr Team, and we run heavily
parallelized unit tests so we enjoy more concurrent VMs.  In case there
are no task processed, the VM count anyways goes down to the setup in
pool.yaml.  So this change actually doesn't mean more VMs is going to be
wasted in normal situations.