releng-compose / compose-rawhide: add mount for ftp archive volume
We are going to sync the contents from fedora_ftp/pub/archive over to
fedora_ftp_archive volume. This will free up 17TB or so on the SSD
aggregate that fedora_ftp is on.

This will mean more mirror churn when we archive old releases, but
there's not much else we can do besides this or more ssd storage.

This is ok to do during freeze as compose-rawhide is not frozen. :)

Signed-off-by: Kevin Fenzi <kevin@scrye.com>