Freeze break request: move /pub/archive from fedora_ftp to fedora_ftp_archive volume
Our fedora_ftp volume is on an SSD aggregate thats running out of space.
So, lets move /pub/archive (17TB) off it on to it's own volume on a
SAS aggregate. archive gets less traffic that other releases, so it
shouldn't be a problem. This will mean however when we archive a release
it will cause a bunch of deletes and re-downloads for mirrors because we
can no longer hardlink content over and then delete it, but there is no
help for that.

I will also notify mirror-admins list about this pending action.
There shouldn't be any short term issues.

Once this PR is merged, we need to run playbooks, then go to a host with
rw access to fedora_ftp and rm the archive tree on it.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>