base / ifcfg: make sure and set staging to use stg.iad2.fedoraproject.org domain
This should only affect stg hosts.
We had set all of iad2 the same, prod and stg both.
We need to make sure stg resolves to stg hosts first.
This worked somewhat until now because we replace the resolv.conf on stg
hosts, but without this they are borken right after boot and until we
replace the resolv.conf and restart httpd or other services.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>