Add a pointer and note about contributions being under the FPCA
Before the ansible repo was public, everyone who could commit to it was
in the sysadmin group, which required FPCA agreement. Now that it's
public we should make sure to clarify this for all users who might
submit contributions. Additionally we should add a note to the pull
request template.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>