Add pipewire and iptables to multilib_whitelist for x86_64
This should fix: https://bugzilla.redhat.com/show_bug.cgi?id=1953178
and pipewire upgrade issues.

Right now the base f34 repo has iptables.i686 and pipewire.i686 in it.

pipewire.i686 would be pulled into updates by mutter.i686, but there's
not yet a mutter update.
iptables.i686 wouldn't be pulled in as nothing depends on it, but it's in the
base repo so thus on users systems.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>