rsyncd / download: adjust odcs to allow package rsyncs
odcs just links to the packages in a compose on the /mnt/koji volume
with relative symlinks. ie (
../../../../../../../../mnt/koji/packages/zip/3.0/28.eln109/data/signed/9867c58f/x86_64/zip-3.0-28.eln109.x86_64.rpm
)
Change odcs on download servers to mount on /mnt like koji and then tell
rsync to chroot at /mnt, so it can follow those links and copy the file
from koji volume.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>