greenwave: specify exact versions for openQA gating policy
I was assuming Bodhi queried on "fedora-rawhide" for Rawhide
updates, but it seems it does not, it queries on "fedora-35" (or
whatever the release number is). We do not want to apply the
gating to Rawhide updates ATM, as openQA does not test them.
So unfortunately we'll need to hand-tend this list unless I can
think of a better approach.

Signed-off-by: Adam Williamson <awilliam@redhat.com>