mirrormanager: install scan-primary-mirror
scan-primary-mirror is another MirrorManager2 tool rewritten in Rust.

The main advantage is that the path to repository mapping is no longer
part of the code, but it is part of the configuration file which makes
it much easier to add new repository definition.

As with all other rewritten tools in Rust it uses considerably less
memory.

Signed-off-by: Adrian Reber <adrian@lisas.de>