copr-be: try to ping6 to wake up ipv6 stack early
Seems like some ppc64le boxes start and seem fine, but they are not
accessible over ipv6 for a long time so the spawner timeouts.  When I
tried to manually ping6 from the box outside - the connection from the
outside started to work immediately - so give it an automatic try.