distgit: Add pull_request_close ACL
Currently it's not possible to create token with pull_request_close ACL
for user (see https://pagure.io/pagure-dist-git/issue/144).

This commit will allow users to add pull_request_close ACL to their
token. The user is still validated if it has the permission, so adding
this ACL to user token doesn't allow user to use API to close any PR
currently opened in dist-git.

Thanks @pingou for helping me with this.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>