add vmhost-x86-09
This host is an old phx2 host thats been sitting there on this ip, but
we haven't moved any vm's to it (thus it shouldn't be affected by the
freeze). However, we should get in ansible so it's network is managed
and so we remember to reboot it for updates, etc.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>