ocp / staging: tell nagios not to check nrpe/swap/mails on ocp hosts
By default nagios assumes it can connect to everything in inventory via
nrpe and monitor things like swap and number of emails in the postfix
queue. For ocp hosts running CoreOS we don't want to have nagios
monitor any of that, we only want it to monitor ping (is the host up).
This change is only in vars here, but it needs a noc run to activate, so
it will need a freeze break to run the noc playbook (as noc is frozen).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>