haproxy: ocp cluster is only in iad2
If we push this to all proxies (as we have), they will fail to start
haproxy because they cannot resolve the internal ocp iad2 hosts. ;(

The ocp clusters should only apply on the iad2 haproxy nodes, not all
proxies. Also fix logic on the staging one to apply in staging instead
of just production.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>