Enable CentOS Stream 9 in MirrorManager
In addition to configuration file changes this also needs the latest
version of MirrorManager, scan-primary-mirror and mirrorlist-server.

The reason for the software updates are mainly that CentOS Stream is
using an empty topdir ('') and all around the code there were
assumptions that topdir always ends with a slash and an empty string
does not end with a slash.

Signed-off-by: Adrian Reber <adrian@lisas.de>