blockerbugs: define BLOCKERBUGS_URL and simplify the config file
BLOCKERBUGS_URL needs to be defined in order to show correct links in the
staging environment. While fixing that, realized that most of the if-else
clauses were unnecessary, because there are already group vars for that. So
simplified the config file by using the vars. KOJI_URL is dropped because it's
no longer needed.

Related: https://pagure.io/fedora-qa/blockerbugs/issue/192