roles/fedoraloveskde: Use hard git resets with an origin
For fedoraloveskde.org, we currently directly push the pre-built website
into the repo branches until we figure out a better approach. Thus the
hisotry does not matter at all here and we just need the latest commit
fetched and switched to. We can thus directly fetch changes and reset,
specify the commit we want to use.

Signed-off-by: Timothée Ravier <tim@siosm.fr>