proxies: redirect old packages to new
The old packages app (which was before that the old community app) lived
on https://apps.fedoraproject.org/packages/ It hasn't been active there
since the datacenter move last year (we retired it). A new openshift app
has been created that lives on https://packages.fedoraproject.org now.
We want to redirect the old path to the new one and remove the old
reverseproxy to nowhere.

Fixes infra ticket #10212

Signed-off-by: Kevin Fenzi <kevin@scrye.com>