batcave: set sftp true because people scp things to batcave
Newer fedora is working on dropping scp (for lots of good reasons)
internally, but using a sftp wrapper. ie, you call scp and internally
ssh uses sftp. This won't work unless the server has sftp enabled.
I (and I suspect other admins) scp things to batcave from time to time
so it would be nice to have this on.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>