Revert "websites: update getfedora.org translation file"
It seems like the pybabel version used on sundries does not
play well with jinja2 files used on getfedora.
Translated versions of this site are impacted as reported by
fedora-web/websites#202

For now, I'll update the pot file manually using a more up to date
version of pybabel until we find a long term solution.

This reverts commit 25fc76f6564befbb1685da990e12c4d03e72a1a2.