simple-koji-ci: retire
simple-koji-ci (currently) is a aws instance that listens to the
fedora-messaging bus and when a PR is filed on src.fedoraproject.org it
takes that PR, applies it to the package git repo and does a scratch
build. It then reports this back to the PR.

However, now we have zuul and fedora-ci doing this (and much more).
simple-koji-ci is a fedora-30 instanced (yes, way eol!)
It's token to talk to src.fedoraproject.org expired long ago and no one
really noticed, so it's not been reporting anything at all for a long
time.

So, lets remove it and let it rest in peace.

Thanks simple-koji-ci! You did great!

Signed-off-by: Kevin Fenzi <kevin@scrye.com>