iddev: retire
iddev was a instance running ipsilon, with a fas plugin. It was usefull
for developers to point to when developing new apps to be able to test
authentication. However, when we moved to new account system the only
way we could make iddev work is by adding it to the VPN and enrolling it
in IPA like a normal machine. We don't want to do that as it would allow
way too much power to anyone who had configuration power over it.

In addition, ryan wrote a vagrant setup called 'tiny-stage' that spins
up a ipa server and a ipsilon server and developers should be able to
use that locally to test, completely disconnected from our
authentication.

See https://pagure.io/fedora-infrastructure/issue/9966 for more info.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>