proxies: fix ETag handling for pkgdb json file
We maintain a static json file for gnome-software (and others) that
tells them when new releases are available. This file gets downloaded a
LOT. Right now because this file is spread out on all our proxies and
the default ETag uses MTime in it, they all return different ETag
values, making it impossible to use ETags to not have to re-download the
file. Setting it to use just Size and Digest for this file means all the
proxies will have the same ETag and gnome-software can see that and not
download the file over and over if it's not changed.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>