sssd / fedorapeople: override homedir in case it gets set to /home
I'm not sure what is causing some users to get /home/user instead of
/home/fedora/user. The ipa default is /home/fedora. Noggin doesn't
appear to override anything.

In any case fedorapeople is where this is really important as we have
web config and the like expecting /home/fedora/user. So, lets override
in sssd to make sure it uses the right place.

I'm also going to manually fix those users that are currently wrong.
See infra 10315

Signed-off-by: Kevin Fenzi <kevin@scrye.com>