copr: define services_disabled per instance
AFAIK we use `services_disabled: true` only when upgrading our severs
from Fedora N to Fedora N+1 which is a big task and we go instances
one by one. Maybe the easier ones in parallel but copr-backend
requires full attention for sure.

If we set `services_disabled: true` for the whole devl/production, we
can't test if the services on each server are starting properly nor we
can't run at least the frontend until all servers are ready.

Let's configure the variable per instance, so whoever is upgrading it
can disable/enable the services when needed.