osbs: exclude docker* from upgrades
Right now we are running a special docker on osbs nodes that allows it
to actually work with f35+ containers. Without this glibc does a syscall
that docker doesn't understand and just blocks, breaking (at least) dns
resolution in the container. So, until we move these nodes from rhel7,
we are going to have to deal with this.

In addtion to excluding this, if it ever gets mistakenly upgraded, you
need to downgrade and then: remove
'--seccomp-profile=/etc/docker/seccomp.json \' from
/usr/lib/systemd/system/docker.service
do 'systemctl daemon-reload'
do 'systemctl restart docker'

Signed-off-by: Kevin Fenzi <kevin@scrye.com>