copr-be: use cronolog also for error.log
I haven't realized this can actually be done (mod_accesslog supports
error.log, too).  And this finally _should be_ the working solution
for now.  At least till we rework the hitcounter entirely, to also
support the AWS CloudFronts logs:

    https://pagure.io/copr/copr/issue/1263

This will allow us to never reload the Lighty server processes for the
log rotation purposes, which turned out to be very problematic for no
obvious reason.  Simply, when the Lighty server is under certain
"production" load (not reproducible via /bin/ab), Lighty fails to reload
(both on SIGHUP and SIGUSR1 signals). Something simply hangs the
processes.

If I had to guess, writes to the pipe to the cronolog process are
blocked causing some weird deadlock?  Since we still have to SIGHUP the
cronolog process, Lighty fails to handle both (a) SIGHUP/SIGUSR1 and (b)
detect cronolog exitted at the same time?  But I'm tired of the
debugging this now.