openqa/server: clean up and trim package requirements
Several of these requirements are old ones that were only needed
for createhdds, when we ran createhdds on the servers. All of
those can go. Also make the list line-by-line for easier git
blame tracking in future (and add comments for the remaining
entries so we know why they're there).

Signed-off-by: Adam Williamson <awilliam@redhat.com>