Use centos+epel-8 for epel-8
We want to move to mock-core-configs-36.4 (pushing to infra 34 repos)
because the version contains multiple config fixes.

That version though dropped epel-8 configs as they go soon EOL.  We plan
to move to rhel+epel, but thad needs more work and testing - so for now
default to centos+epel again (CentOS 8 goes EOL in Jan 2022).