httpd: redirect copr /api/ URLs to fedorainfracloud.org
See https://pagure.io/copr/copr/issue/1935

Prevent only /api_2, /api_3, etc from redirectring from
fedoraproject.org to fedorainfracloud.org

I am not entirely sure why do we need these special-cases for API but
you guys are taking care of it and keeping it updated, so it must be
important. If anyone can explain some context, it would be
appreciated. The config is older than my involvement in the Copr
project :-)

Anyway, there is only one /api/ page - https://copr.fedorainfracloud.org/api
and that is not an API endpoint that is programmatically accessed and
that preserves backwards compatibility. It is a page that one opens
in the web browser to find information about API, such as where the
documentation is, and how to obtain an API token.

We would like to apply the redirect from fedoraproject.org to
fedorainfracloud.org even for this page.