copr-builders: other s390x IBM Cloud fixes
Don't install subscription-manager when it is already installed.  Don't
even try to download the DNF metadata for that (at that point, the
subscription is not yet activated, and so the attempt to install
subscription-manager subjectively slows-down the spawning).

Try to subscribe a bit sooner, so we don't wait till the end of the
playbook for a potential subscription failure.

Run subscription-manager clean to which seems to way-around the
katello-* RPM hack.

One forgotten ibm-cloud-vm typofix (name vs. opts.name).

Define server_hostname to enforce the default rhsm server.