buildvm_s390x: adjust mem and cpus on kvm s390x builders
We got some more resources a while back and so we wanted to make more
builders, but now we are waiting on disk space to do that. In the mean
time and since the mass rebuild is coming up tuesday, just upgrade them
to f35 and use the memory and cpus we have.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>