Try to restart a mirrorlist-server process if not running
Sometimes we see that mirrorlist-server processes are not running and it
just needs to be restarted. Let the script try it twice before erroring
out.

The most common reason for non running mirrorlist-server processes seems
to be incomplete or truncated input files which is usually fixed the
next time the data is synced.

Signed-off-by: Adrian Reber <adrian@lisas.de>