buildvm_s390x: reorg group and host vars, split zvm/kvm lpars
We can consolidate a bunch of things here in group vars that was in
host vars and avoid a bunch of duplication. So, now: all the
buildvm_s390x builders are in that group, but there's now a
buildvm_s390x_kvm and buildvm_s390x_zvm for those seperate lpars and
those define the default network setup for each lpar. Finally the hosts
vars only now needs to have the ip addresses.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>