copr/certbot: revert the automatic backup/restore of certs
For synchronizing from host A to host B we would have to have SSH
configured from A to B or from B to A (which we don't, and we don't even
want to).  The slurp solution would work, but certbot is pretty picky
WRT the restored /etc/letsencrypt directory (structure, filenames,
symlinks..).