Greenwave: separate upgrade tests in openQA update gating policy
The upgrade tests are not run on updates for the oldest stable
release, because we don't want to maintain base images for EOL
releases. So we can't gate the oldest stable release (currently
F34) on those tests as we don't run them. To handle this let's
split the policy in two and have the policy for the upgrade tests
apply only to the release(s) they're actually run on.

Signed-off-by: Adam Williamson <awilliam@redhat.com>