coreos-ci: rename anyuid-setfcap SCC
SCCs are cluster-scoped, but here we just want an SCC uniquely for use
by the `coreos-ci` project. So prefix the SCC name with that.

This may have been the root of an issue where we had this SCC defined
mulitple times (once here and once in the `fedora-coreos-pipeline` role)
and any users added were lost because it was being redefined.

The one in `fedora-coreos-pipeline` has since been nuked:
https://pagure.io/fedora-infra/ansible/pull-request/989