drop local clevis role and replace with linux-system-roles.nbde_client
The linux-system-roles.nbde_client does the right things and so one less
thing for us to maintain. It also avoids the problem we have on some
machines now where network interface names are not as the old clevis
role expected, so unlocking didn't work on boot. Hopefully this fixes
all those issues.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>