nbde: fix this with proper fixes
My last commit wasn't correct. It turns out we still need to:
a) drop the dracut made bogus "Wired Connection". We do this in
nbde_client-network-flush by removing the default connection file.
b) We also need to tell NetworkManager to not make 'Wired Connection 43'
for any non configured interfaces. We only want to use the ones we
configure and not any of the others that may or may not have link.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>