ansible-server: Use RPM collection packages
We now have all the collections that are used by Fedora Infra packaged
for EPEL 8 (the ansible control host runs on RHEL 8, as far as I know),
except community.postgresql. ansible-collection-community-rabbitmq[1]
and ansible-collection-community-libvirt[2] are new packages that are
still in epel8-next testing, so I would wait to merge this until they're
pushed.

I would also recommend clearing the current collections directory (`rm
-rf ~/.ansible/ansible_collections/*`) before merging this to ensure
that ansible uses the packaged versions and not the old versions that
were installed with ansible-galaxy.

[1]: https://bodhi.fedoraproject.org/updates/FEDORA-EPEL-2022-a7237cc021
[2]: https://bodhi.fedoraproject.org/updates/FEDORA-EPEL-2022-a9ccbc0012

Signed-off-by: Maxwell G <gotmax@e.email>