inventory: try and setup iad2_production and iad2_staging groups
This is needed because I want to override just the mtu setting from
group_vars/all, but I don't want to copy the network_connections block
to every host in iad2 to override it. Hopefully this will allow us to
override it on just iad2_production (staging is already actually set in
group_vars/staging, but for completeness we may want a staging group
someday here, so I went and added it.).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>