proxies: drop ipv6 restart script
Yesterdays fix: a script that restarted networking to bring back a ipv6
route on proxy11, causes:
Todays bug: restarting the interface every minute causes the nameserver
to burp and not find sundries01, so all the rsyncs fail. It seems to
handle this just fine now, so lets retire this glorious hack.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>